backend/requirements.txt
asgiref==3.4.1
astroid==2.6.2
certifi==2021.5.30
cffi==1.14.5
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
docusign-esign==3.10.0
gunicorn==20.1.0
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
nose==1.3.7
pycparser==2.20
PyJWT==1.7.1
pylint==2.9.3
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
sqlparse>=0.4.2
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3>=1.26.5
wrapt==1.12.1
